app/authentication/models.py
from tortoise import fields
from tortoise.models import Model


class User(Model):
    id = fields.IntField(pk=True)
    email = fields.CharField(max_length=255, unique=True)
    password = fields.CharField(max_length=255)
    external_id = fields.IntField(unique=True)

    class Meta:
        table = "users"

