requirements/base.txt
fastapi==0.95.2
uvicorn==0.23.2
gunicorn
python-multipart==0.0.9
pypdf==4.2.0
aiohttp==3.11.13
tortoise-orm==0.20.0
python-jose
python-dotenv==1.0.0
asyncpg
aerich==0.7.1
pydantic==1.10.13
